Added heat map to analysis pipeline
